added ethans, nicks, melodies parts to push
