+++ client/addressfilter.html
@@ -1,0 +1,55 @@
+<!-- ethan ic3 html-->
+<html>
+<head>
+    <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>CP Restaurant Database</title>
+    <link rel="stylesheet" href="style/styles.css" />
+</head>
+
+<body>
+    <h1 class="header1">College Park Food Finder</h1>
+    <nav>
+        <a href="index.html">Home</a>
+        <a href="about.html">About</a>
+        <a href="addRestaurant.html">Add Restaurant</a>
+        <a href="filters.html">Filters</a>
+        <a href="restaurants.html">Restaurants</a>
+        <a href="addressfilter.html">Search Restaurant</a>
+    </nav>
+    <div>
+        <form class="mainform">
+            <div class="field">
+                <label>
+                    Address 1
+                </label>
+                <input type="text" name="Address_1" id="NewAddress">
+            </div>
+            <div class="field">
+                <label>
+                    Address_2
+                </label>
+                <input type="text" name="Address_2" id="Address_2">
+            </div>
+            <div class="control">
+                <button class="button">
+                    submit
+                </button>
+            </div>
+        </form>
+    </div>
+
+    <div class="restaurant_name">
+        <h2>Restaurants</h2>
+        <ul class="resto-list">
+
+        </ul>
+    </div>
+    <script src="loadedscriptEthan.js">
+
+    </script>
+</body>
+
+</html>
+
