--- client/combined.html
@@ -1,116 +1,0 @@
-<!-- melody ic3 html-->
-<html>
-
-<head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>CP Restaurant Database</title>
-    <link rel="stylesheet" href="style/styles.css"/>
-</head>
-<body>
-    <h1 class="header1">College Park Food Finder</h1>
-    <nav>
-        <a href="/index.html">Home</a>
-        <a href="/about.html">About</a>
-        <a href="/addRestaurant.html">Add Restaurant</a>
-        <a href="/filters.html">Filters</a>
-        <a href="/restaurants.html">Restaurants</a>
-
-    </nav>
-    <div>
-        <form class ="mainform">
-            <div class = "field">
-                <label>
-                    Takeout 
-                </label>
-                <input type="checkbox" name="takeout" id="takeout">
-            </div>
-            <div class = "field">
-                <label>
-                    Delivery 
-                </label>
-                <input type="checkbox" name="delivery" id="delivery">
-            </div>
-            <div class = "field">
-                <label>
-                    Parking
-                </label>
-                <input type="checkbox" name="parking" id="parking">
-            </div>
-            <div class = "control">
-                <button class = "button">
-                    submit
-                </button>
-            </div>
-        </form>
-        <table style="margin: 100px;background-color: aliceblue" class="table">
-            <!--First tr is the column headers-->
-            <tr>
-                <th>Restaurant Name</th>
-                <th>Description</th>
-            
-            </tr>
-        </table>
-    </div>
-    <script src = "loadtable.js">
-
-    </script>
-</body>
-</html>
-
-<!-- ethan ic3 html-->
-<html>
-
-<head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>CP Restaurant Database</title>
-    <link rel="stylesheet" href="style/styles.css" />
-</head>
-
-<body>
-    <h1 class="header1">College Park Food Finder</h1>
-    <nav>
-        <a href="/index.html">Home</a>
-        <a href="/about.html">About</a>
-        <a href="/addRestaurant.html">Add Restaurant</a>
-        <a href="/filters.html">Filters</a>
-        <a href="/restaurants.html">Restaurants</a>
-    </nav>
-    <div>
-        <form class="mainform">
-            <div class="field">
-                <label>
-                    Address 1
-                </label>
-                <input type="text" name="Address_1" id="NewAddress">
-            </div>
-            <div class="field">
-                <label>
-                    Address_2
-                </label>
-                <input type="text" name="Address_2" id="Address_2">
-            </div>
-            <div class="control">
-                <button class="button">
-                    submit
-                </button>
-            </div>
-        </form>
-    </div>
-
-    <div class="restaurant_name">
-        <h2>Restaurants</h2>
-        <ul class="resto-list">
-
-        </ul>
-    </div>
-    <script src="loadedscriptEthan.js">
-
-    </script>
-</body>
-
-</html>
-
