+++ client/index.html
@@ -21,6 +21,8 @@
         <a href="addRestaurant.html">Add Restaurant</a>
         <a href="filters.html">Filters</a>
         <a href="restaurants.html">Restaurants</a>
+        <a href="addressfilter.html">Search Restaurant</a>
+
     </nav>
 
     <div>
