+++ server.js
@@ -4,6 +4,8 @@ import db from './database/initializeDB.js';
 import apiRoutes from './routes/apiRoutes.js';
 import apiRoutesJessica from './routes/apiRoutesJessica.js';
 import apiRoutesMelody from './routes/apiRoutesMelody.js';
+import apiRoutesEthan from './routes/apiRoutesEthan.js';
+import apiRoutesNick from './routes/apiRoutesNick.js';
 
 
 const app = express();
@@ -17,7 +19,7 @@ app.use(express.json());
 app.use(express.static(staticFolder));
 app.use('/api', apiRoutes);
 app.use('/jess', apiRoutesJessica);
-
+app.use('/ethan', apiRoutesEthan);
 app.use('/melody', apiRoutesMelody);
